Cache reflective calls
The prior implementation of the packet manager would run the respective
reflective constructor/method/field lookups when the logic that required
these was executed.
Besides the lazy loading of these reflective references, the results
were not cached anywhere and hence re-computed every time any method on
the packet manager was called.

To improve both speed and error tracing, this commit moves all the
reflective initialization logic into a newly introduced factory for each
packet manager. This factory will locate any constructor/method/field
needed for the packet manager of a specific version and pass the
instances to the constructor of that packet manager which then stores
them as fields.

This change allows the packet managers to directly access the reflection
instances needed to access server internals as well as ensures that any
reflection error is cought at creation time of the packet manager
instead of runtime.