Show indicator on arrow poison damage
This commit adds functionality that detects a hit dealt by an arrow of poison and shows indicators for damage dealt by that arrows poison effect.
To display the indicator to the right player, the damagers UUID is saved on the damaged entity and is checked for on every tick a mob gets poison damage. The UUID is removed after the poison effect duration.
Due to the way this is solved, after the entity is poisoned by another player, that player will take over 'damage credit' and the indicator will only be shown to the last player who applied poison. This is effect is desired if we don't consider the upgraded version of poison, 'poison II'. In minecraft poison II takes priority, where as here if the entity already has poison II applied and another player uses a Poison I arrow, they will still take over the credit even though the entities poison effect stays untouched.